move HostAppointer into its own module
